Save mobile number after verifying it
In the current implementation when multi step authentication is enabled and if the user has not specify the mobile number in the profile, it will prompt user to enter mobile number and then directly saves it. If user enters an
incorrect mobile there is no way for the user to change it himself. This implementation will save the mobile number once it is confirmed with the OTP code sends.